svelte-migrate: update files
