+++ src/routes/blog/+page.svelte
@@ -1,15 +1,4 @@
-<script lang="ts" context="module">
-    export const load = async ({fetch}) => {
-        const posts = await fetch('/api/posts.json')
-        const allPosts = await posts.json()
 
-        return {
-            props: {
-                posts: allPosts
-            }
-        }
-    }
-</script>
 
 <script>
     export let posts
