+++ src/routes/blog/+page.ts
@@ -1,0 +1,9 @@
+
+export const load = async ({fetch}) => {
+    const posts = await fetch('/api/posts.json')
+    const allPosts = await posts.json()
+
+    return {
+    posts: allPosts
+}
+}
