+++ src/routes/projects/+page.svelte
@@ -1,15 +1,4 @@
-<script lang="ts" context="module">
-    export const load = async ({fetch}) => {
-        const projects = await fetch('/api/projects.json')
-        const allProjects = await projects.json()
 
-        return {
-            props: {
-                projects: allProjects
-            }
-        }
-    }
-</script>
 
 <script>
     export let projects
