+++ src/routes/projects/projectLayout/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import Layout from './_project.svelte'
+	import Layout from '../_project.svelte'
     export let title
     export let status
     export let desc
